requirements.txt
streamlit
python-dotenv
groq==0.11.0
httpx==0.27.0